chore: fmt

+++ nova_vm/src/ecmascript/builtins/text_processing/string_objects/string_prototype.rs
@@ -1038,7 +1038,9 @@ fn string_pad(
         let mut vec = VecDeque::with_capacity(2);
         let mut sub_string = vec![0; fill_len as usize];
         for i in 0..fill_len {
-            fill_string.utf16_char(agent, i as usize).encode_utf8(&mut sub_string[fill_string.utf8_index(agent, i as usize).unwrap() as usize..]);
+            fill_string.utf16_char(agent, i as usize).encode_utf8(
+                &mut sub_string[fill_string.utf8_index(agent, i as usize).unwrap()..],
+            );
         }
         let sub_string = std::str::from_utf8(&sub_string).unwrap();
         // let sub_string = &fill_string.as_str(agent)[..fill_len as usize];
@@ -1051,7 +1053,9 @@ fn string_pad(
         let sub_string_len = (fill_len % fill_string_len) as usize;
         let mut sub_string = vec![0; sub_string_len];
         for i in 0..sub_string_len {
-            fill_string.utf16_char(agent, i).encode_utf8(&mut sub_string[fill_string.utf8_index(agent, i).unwrap() as usize..]);
+            fill_string
+                .utf16_char(agent, i)
+                .encode_utf8(&mut sub_string[fill_string.utf8_index(agent, i).unwrap()..]);
         }
         let sub_string = std::str::from_utf8(&sub_string).unwrap();
         vec.push_back(String::from_string(agent, sub_string.to_owned()));
